Corrected CaseSensitiveFilename and found memory error with valgrind ...
... and a bit of luck:

When trying to get NValues in gdb, it wasn't found in the context (actually
seems to have been a gdb weird error or misuse). When trying to find out what
NValues was, I found the error: ++dataInserted was done before actually
inserting the data, which lead to the last data pointing to unallocated memory!

Well another 4+ hours lost, but I'm beginning to become pro with Linux and
the 'toolchain'. I always wondered what programmers meant with 'toolchain',
now I more or less understand.